+++ .php-cs-fixer.fixture.php
@@ -37,6 +37,7 @@ $config = PhpCsFixer\Config\Factory::fromRuleSet($ruleSet);
 $config->getFinder()
     ->in(__DIR__ . '/test/Fixture/')
     ->notPath([
+        'CallLikes/NoNamedArgumentRule/script.php',
         'Closures/NoNullableReturnTypeDeclarationRule/script.php',
         'Closures/NoParameterWithNullableTypeDeclarationRule/script.php',
         'Functions/NoNullableReturnTypeDeclarationRule/script.php',