+++ test/Fixture/CallLikes/NoNamedArgumentRule/AbstractClass.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Ergebnis\PHPStan\Rules\Test\Fixture\CallLikes\NoNamedArgumentRule;
+
+abstract class AbstractClass
+{
+    abstract public function __construct($bar = null);
+
+    abstract public function bar($baz, $qux): void;
+
+    abstract public static function create($bar = null): self;
+}