+++ test/Fixture/CallLikes/NoNamedArgumentRule/ExampleClassExtendingAbstractClassAndImplementingExampleInterface.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Ergebnis\PHPStan\Rules\Test\Fixture\CallLikes\NoNamedArgumentRule;
+
+class ExampleClassExtendingAbstractClassAndImplementingExampleInterface extends AbstractClass implements ExampleInterface
+{
+    public function __construct($bar = null)
+    {
+    }
+
+    public function bar($baz, $qux): void
+    {
+    }
+
+    public static function create($bar = null): self
+    {
+        return new self($bar);
+    }
+}